pyproject.toml
[project]
name = "rag_agent_com_reranking"
version = "0.0.1"

requires-python = ">=3.9"
readme = "README.md"

dependencies = [
    "python-dotenv>=0.20.0",
    "openai>=1.10.0",
    "cohere>=5.0.0",
    "numpy>=1.23.0",
    "tqdm>=4.60.0",
    "langchain-core>=0.1.40",
    "langchain-openai>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "langgraph>=0.0.30",
    "tiktoken>=0.5.0",
    "ipython>=7.0.0"
]
